requirements.txt
Pillow >= 7.0.0
qrcode >= 6.1
lxml >= 4.5.0
PyYAML >= 5.3
